+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Changes can be:
 
 * 💫 Redesign examples viewer to be more readable and in a way that doesn't force `display: flex` onto contents.
 
+* 💫 Introduce client-side routing for static builds to make page transitions smoother by default. In addition, the option `:bundle?` for `clerk/build!` is now deprecated in favour of setting a new option `:package` to `:single-file` (the default for it being `:directory`).
+
 * 🛠 Bump depdendencies
 
   * `com.taoensso/nippy` to `3.4.0-beta1`